refactor: [US-045] - Move SEO views into core/views/seo.py
